# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Helix server integration API",
    "version": "1.0.0",
    "description": "Private-LAN server management API. Session cookie AND CSRF proof are required together; mutations and file commands require exact Origin. Account capabilities are not per-server credentials. All server routes are listed; broker response objects are not yet all fully typed. Consult per-server capabilities before acting. Do not automatically replay mutations or treat a job ID as completion."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "SessionCookie": [],
      "CsrfProof": []
    }
  ],
...
  "paths": {
...
    "/auth/server-tokens": {
      "get": {"operationId":"listServerTokens","summary":"List token metadata (expires_at is null for never-expiring tokens; authorized is false after an owner account change; viewable is true when an encrypted copy can be shown again) and permission names; never returns secrets","responses":{"default":{"$ref":"#/components/responses/Problem"},"200":{"description":"Token catalog"}}},
      "post": {"operationId":"createServerToken","summary":"Issue a scoped server token with a 1–365 day or no expiration (null); owner session only. The permission all grants every server permission, including later ones, but never host controls","parameters":[{"$ref":"#/components/parameters/Origin"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["name","servers","permissions","expires_in_days"],"properties":{"name":{"type":"string","minLength":1,"maxLength":80},"servers":{"type":"array","minItems":1,"maxItems":64,"items":{"type":"string"}},"permissions":{"type":"array","minItems":1,"items":{"type":"string","enum":["view","logs","files.read","files.write","start","stop","restart","kill","console","settings","update","backups.read","backups.write","network","remove","all"]}},"expires_in_days":{"oneOf":[{"type":"integer","minimum":1,"maximum":365},{"type":"null"}]}}}}}},"responses":{"default":{"$ref":"#/components/responses/Problem"},"200":{"description":"Token secret; viewable says whether the owner can show it again later"},"403":{"description":"Invalid scope or insufficient owner permissions"}}}
    },
    "/auth/server-tokens/{id}": {
      "delete":{"operationId":"revokeServerToken","summary":"Revoke a token without cancelling accepted jobs","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"$ref":"#/components/parameters/Origin"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","maxProperties":0}}}},"responses":{"default":{"$ref":"#/components/responses/Problem"},"200":{"description":"Revoked"},"404":{"description":"Unknown token"}}}
    },
    "/auth/server-tokens/{id}/rotate": {
      "post":{"operationId":"rotateServerToken","summary":"Replace a live token secret while preserving its scope and expiration","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"$ref":"#/components/parameters/Origin"}],"responses":{"default":{"$ref":"#/components/responses/Problem"},"200":{"description":"New token secret; old secret invalidated and its stored copy deleted"},"404":{"description":"Unknown, expired or revoked token"}}}
    },
    "/auth/server-tokens/{id}/reveal": {
      "post":{"operationId":"revealServerToken","summary":"Show a stored token secret again after the owner re-enters their dashboard password; audited","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"$ref":"#/components/parameters/Origin"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["current_password"],"properties":{"current_password":{"type":"string"}}}}}},"responses":{"default":{"$ref":"#/components/responses/Problem"},"200":{"description":"Token secret"},"401":{"description":"Password rejected (current_password_rejected) or session expired"},"404":{"description":"Unknown or revoked token, or no stored copy"}}}
    },
    "/automation/server": {
      "post":{"operationId":"executeScopedServerRequest","summary":"Run a permitted operation on a token-scoped server","description":"Bearer token only. Cookie and Origin headers are rejected. See SERVER-TOKENS.md for operation fields and permissions. No host/global operations are allowed; job_status is restricted to this token's accepted jobs.","security":[{"ServerToken":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["operation"],"properties":{"operation":{"type":"string","enum":["server_capabilities","server_detail","server_logs","server_log_history","server_files","server_action","server_console","server_settings","update_server_settings","set_native_memory","set_native_cpu","set_native_start_on_boot","set_native_browser_listing","change_native_runtime","list_backups","server_backup_download","restore_backup","trash_backup","restore_trashed_backup","set_backup_policy","prune_backups","job_status"]},"instance_id":{"type":"string"},"job_id":{"type":"string"}}},"examples":{"stop":{"value":{"operation":"server_action","instance_id":"helix:EXACT-UUID","action":"stop"}},"files":{"value":{"operation":"server_files","instance_id":"helix:EXACT-UUID","request":{"action":"list","path":"","limit":50}}}}}}},"responses":{"default":{"$ref":"#/components/responses/Problem"},"200":{"description":"Operation response or job ID; check job completion separately"},"401":{"description":"Invalid, expired or revoked token"},"403":{"description":"Server, operation or permission is outside token scope"},"429":{"description":"Token request rate exceeded"}}}
    },
    "/automation/jobs": {
      "get":{"operationId":"listScopedServerJobs","summary":"List the last 256 jobs accepted through this token","security":[{"ServerToken":[]}],"responses":{"default":{"$ref":"#/components/responses/Problem"},"200":{"description":"Token-owned job IDs and server IDs"},"401":{"description":"Invalid token"}}}
    },
...
    "/servers": {
      "get": {
        "operationId": "listServers",
        "tags": [
          "Servers"
        ],
        "x-helix-capability": "games.view",
        "summary": "Discover server identities; never target a server by list order or version",
        "description": "Use each item's id unchanged in the instance_id path parameter. Imported servers belong to their external manager. Names may be duplicated. Check manager/readiness and server detail before writing. Do not infer permissions from the presence of an item.",
        "responses": {
          "200": {
            "description": "Server inventory array",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Server"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "parameters": []
    },
...
    "/servers/{instance_id}": {
      "parameters": [
        {
          "name": "instance_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "getServer",
        "tags": [
          "Servers"
        ],
        "x-helix-capability": "games.view",
        "summary": "Read a native server; imported servers use inventory and capabilities",
        "responses": {
          "200": {
            "$ref": "#/components/responses/BrokerObject"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ServerToken": {"type":"http","scheme":"bearer","description":"Server-scoped token; accepted only by /automation endpoints."},
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "helix_session"
      },
      "CsrfProof": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Helix-CSRF"
      }
    },
    "parameters": {
      "Origin": {
        "name": "Origin",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uri"
        },
        "description": "Exact configured Helix origin: scheme, host and port, without a trailing slash. Never spoof a different service's origin."
      },
...
    },
    "responses": {
      "Problem": {
        "description": "Error; branch on HTTP status and code, not message text. A mutation's outcome may be unknown after a timeout or transport error.",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Retry-After": {
            "schema": {
              "type": "string"
            },
            "description": "Present on selected capacity/rate errors. Does not make a mutation safe to replay."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "BrokerObject": {
        "description": "Broker result. The exact resource-specific fields remain documented in docs/API.md and tested in frontend adapters; this schema deliberately does not claim full generated-client coverage.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "schemas": {
...
      "Server": {
        "type": "object",
        "required": [
          "id",
          "name",
          "instance_name",
          "manager",
          "software",
          "version",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "instance_name": {
            "type": "string"
          },
          "manager": {
            "type": "string"
          },
          "software": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "game_port": {
            "type": [
              "integer",
              "null"
            ]
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
    }
  }
}